start restructuring program
