+++ programs/gateway/src/instructions/mod.rs
@@ -1,0 +1,5 @@
+// Re-export all instruction handlers from submodules
+pub mod admin;
+pub mod deposit;
+pub mod withdraw;
+pub mod execute;
